+++ src/scopes/SearchQueryBase.cpp
@@ -64,6 +64,13 @@ QueryCtrlProxy SearchQueryBase::subsearch(ScopeProxy const& scope,
     return subsearch(scope, query_string, "", FilterState(), fwd()->search_metadata(), reply);
 }
 
+QueryCtrlProxy SearchQueryBase::subsearch(ChildScope const& scope,
+                                          string const& query_string,
+                                          SearchListenerBase::SPtr const& reply)
+{
+    return subsearch(scope, query_string, "", FilterState(), fwd()->search_metadata(), reply);
+}
+
 QueryCtrlProxy SearchQueryBase::subsearch(ScopeProxy const& scope,
                                           std::string const& query_string,
                                           FilterState const& filter_state,
@@ -81,6 +88,15 @@ QueryCtrlProxy SearchQueryBase::subsearch(ScopeProxy const& scope,
     return subsearch(scope, query_string, department_id, filter_state, fwd()->search_metadata(), reply);
 }
 
+QueryCtrlProxy SearchQueryBase::subsearch(ChildScope const& scope,
+                                          std::string const& query_string,
+                                          std::string const& department_id,
+                                          FilterState const& filter_state,
+                                          SearchListenerBase::SPtr const& reply)
+{
+    return subsearch(scope, query_string, department_id, filter_state, fwd()->search_metadata(), reply);
+}
+
 QueryCtrlProxy SearchQueryBase::subsearch(ScopeProxy const& scope,
                                           std::string const& query_string,
                                           std::string const& department_id,
@@ -88,7 +104,17 @@ QueryCtrlProxy SearchQueryBase::subsearch(ScopeProxy const& scope,
                                           SearchMetadata const& hints,
                                           SearchListenerBase::SPtr const& reply)
 {
-    return fwd()->subsearch(scope, query_string, department_id, filter_state, hints, reply);
+    return fwd()->subsearch(scope, {}, query_string, department_id, filter_state, hints, reply);
+}
+
+QueryCtrlProxy SearchQueryBase::subsearch(ChildScope const& scope,
+                                          std::string const& query_string,
+                                          std::string const& department_id,
+                                          FilterState const& filter_state,
+                                          SearchMetadata const& hints,
+                                          SearchListenerBase::SPtr const& reply)
+{
+    return fwd()->subsearch(scope.metadata.proxy(), scope.keywords, query_string, department_id, filter_state, hints, reply);
 }
 
 void SearchQueryBase::set_department_id(std::string const& department_id)