+++ src/scopes/internal/zmq_middleware/ScopeI.cpp
@@ -19,17 +19,18 @@
 #include <unity/scopes/internal/zmq_middleware/ScopeI.h>
 
 #include <scopes/internal/zmq_middleware/capnproto/Scope.capnp.h>
-#include <unity/scopes/internal/ResultImpl.h>
-#include <unity/scopes/internal/CannedQueryImpl.h>
-#include <unity/scopes/internal/ScopeObject.h>
+#include <unity/scopes/CannedQuery.h>
 #include <unity/scopes/internal/ActionMetadataImpl.h>
+#include <unity/scopes/internal/CannedQueryImpl.h>
+#include <unity/scopes/internal/ResultImpl.h>
 #include <unity/scopes/internal/SearchMetadataImpl.h>
+#include <unity/scopes/internal/ScopeMetadataImpl.h>
+#include <unity/scopes/internal/ScopeObject.h>
 #include <unity/scopes/internal/zmq_middleware/ObjectAdapter.h>
 #include <unity/scopes/internal/zmq_middleware/VariantConverter.h>
 #include <unity/scopes/internal/zmq_middleware/ZmqQueryCtrl.h>
 #include <unity/scopes/internal/zmq_middleware/ZmqReply.h>
 #include <unity/scopes/internal/zmq_middleware/ZmqScope.h>
-#include <unity/scopes/CannedQuery.h>
 
 #include <cassert>
 
@@ -201,8 +202,8 @@ void ScopeI::preview_(Current const& current,
 }
 
 void ScopeI::child_scopes_(Current const&,
-                      capnp::AnyPointer::Reader&,
-                      capnproto::Response::Builder& r)
+                           capnp::AnyPointer::Reader&,
+                           capnproto::Response::Builder& r)
 {
     auto delegate = dynamic_pointer_cast<ScopeObjectBase>(del());
     assert(delegate);
@@ -213,18 +214,27 @@ void ScopeI::child_scopes_(Current const&,
     auto list_response = r.initPayload().getAs<capnproto::Scope::ChildScopesResponse>();
     auto list = list_response.initReturnValue(child_scopes.size());
 
-    int i = 0;
-    for (auto const& child_scope : child_scopes)
+    for (size_t i = 0; i < child_scopes.size(); ++i)
     {
-        list[i].setId(child_scope.id);
-        list[i].setEnabled(child_scope.enabled);
-        ++i;
+        list[i].setId(child_scopes[i].id);
+
+        auto dict = list[i].initMetadata();
+        to_value_dict(child_scopes[i].metadata.serialize(), dict);
+
+        list[i].setEnabled(child_scopes[i].enabled);
+
+        auto keywords = list[i].initKeywords(child_scopes[i].keywords.size());
+        int j = 0;
+        for (auto const& kw : child_scopes[i].keywords)
+        {
+            keywords.set(j++, kw);
+        }
     }
 }
 
-void ScopeI::set_child_scopes_(Current const&,
-                      capnp::AnyPointer::Reader& in_params,
-                      capnproto::Response::Builder& r)
+void ScopeI::set_child_scopes_(Current const& current,
+                               capnp::AnyPointer::Reader& in_params,
+                               capnproto::Response::Builder& r)
 {
     auto delegate = std::dynamic_pointer_cast<ScopeObjectBase>(del());
     assert(delegate);
@@ -235,8 +245,22 @@ void ScopeI::set_child_scopes_(Current const&,
     for (size_t i = 0; i < list.size(); ++i)
     {
         string id = list[i].getId();
+
+        auto md = list[i].getMetadata();
+        VariantMap m = to_variant_map(md);
+        unique_ptr<ScopeMetadataImpl> smdi(new ScopeMetadataImpl(m, current.adapter->mw()));
+        auto metadata = ScopeMetadata(ScopeMetadataImpl::create(move(smdi)));
+
         bool enabled = list[i].getEnabled();
-        child_scope_list.push_back( ChildScope{id, enabled} );
+
+        set<string> keywords;
+        auto keywords_capn = list[i].getKeywords();
+        for (auto const& kw : keywords_capn)
+        {
+            keywords.emplace(kw);
+        }
+
+        child_scope_list.push_back( ChildScope{id, metadata, enabled, keywords} );
     }
 
     bool result = delegate->set_child_scopes(child_scope_list);